+++ src/index.tsx
@@ -1,30 +1,10 @@
 import "./instrument";
 import { createRoot } from "react-dom/client";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import DocumentTitle from "components/DocumentTitle";
 import ErrorBoundary from "./components/ErrorBoundary";
-import App from "./App";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import AppWithAtomsCompat from "AppWithAtomsCompat";
 import "./index.scss";
-import { withAtomsCompat } from "states/atomUtils";
-import { userSettingsState } from "./states/userSettingsState";
-import { globalUserSettingsState } from "./states/globalUserSettingsState";
-import { useLessonIndexWithFallback } from "./states/lessonIndexState";
-import { revisionModeState } from "./states/lessonState";
-import Analytics from "components/Analytics";
-
-// eslint-disable-next-line react-refresh/only-export-components
-function AppWrapper(props: object) {
-  const lessonIndex = useLessonIndexWithFallback();
-  // @ts-expect-error TS(2740) FIXME: Type '{ lessonIndex: LessonIndexEntry[]; }' is mis... Remove this comment to see the full error message
-  const app = <App {...props} {...{ lessonIndex }} />;
-  return <Analytics>{app}</Analytics>;
-}
-
-const AppWithAtomsCompat = withAtomsCompat(AppWrapper, [
-  ["revisionMode", revisionModeState],
-  ["userSettings", userSettingsState],
-  ["globalUserSettings", globalUserSettingsState],
-]);
 
 const container = document.getElementById("root");
 const root = createRoot(container!);
@@ -44,5 +24,5 @@ root.render(
         </Routes>
       </ErrorBoundary>
     </Router>
-  </DocumentTitle>
+  </DocumentTitle>,
 );
